Added some enemies, changed floor color and added pickupable screwdriver + saw.
I want to note that there are a good bit more items that can be assessed under "OutsideAssets" and either of the folders -> prefabs. However, these are "raw", meaning that I haven't checked if the collidier is correct, if it has a rigidbody, they aren't pickupable, etc. I was going to change the prefabs of the boxes to make them pickupable, but I wanted to wait for the merger with the new pick-up system, as that changes things.